+++ essencium-backend/src/main/java/de/frachtwerk/essencium/backend/model/dto/EmailVerificationRequest.java
@@ -1,0 +1,6 @@
+package de.frachtwerk.essencium.backend.model.dto;
+
+import jakarta.validation.constraints.NotNull;
+import java.util.UUID;
+
+public record EmailVerificationRequest(@NotNull UUID emailVerifyToken) {}